chore: add .python-version and pyproject.toml files
Add .python-version file to specify the Python version used in the project (3.11.4/envs/scrapy-env).

Add pyproject.toml file to manage project dependencies and metadata. Poetry is used as the dependency management tool. The project name is "rentcrawler" with version 0.1.0. The dependencies include Python 3.10, Scrapy 2.9.0, scrapy-poet 0.13.0, scrapy-redis 0.7.3, and scrapy-fake-useragent 1.4.4. Additionally, pymongo 4.4.1 is added as a development dependency.

Remove requirements.txt file as it is no longer needed.