feat(data): add configuration files for Quinto Andar and Viva Real APIs
Add configuration files for Quinto Andar and Viva Real APIs to the data directory. These files contain the necessary information such as the API endpoints, headers, and request bodies for making requests to these APIs. These configuration files will be used in the application to fetch data from Quinto Andar and Viva Real APIs for further processing.

feat(zap.json): add zap.json file to store API data for ZapImóveis integration

The zap.json file is added to the data directory. This file will be used to store the API data for the ZapImóveis integration. The data object in the file contains the URL and headers for making the API request, and the params object contains the pagination information. This file will be used by the application to fetch and process the data from the ZapImóveis API.